% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggwebgl-adapter.R
\name{as_ggwebgl_spec}
\alias{as_ggwebgl_spec}
\title{Convert an object to a ggWebGL primitive specification}
\usage{
as_ggwebgl_spec(x, ...)
}
\arguments{
\item{x}{Object to convert.}

\item{...}{Additional arguments.}
}
\description{
This generic is defined locally so \code{boids4R} can offer an optional ggWebGL
adapter without depending on ggWebGL at load time.
}
